% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up.R
\name{clean_up}
\alias{clean_up}
\title{Clean up data imported using LCL::pull_data.}
\usage{
clean_up(x)
}
\arguments{
\item{x}{A data frame or tibble containing data pulled using LCL::pull_data}
}
\description{
This function uses the AMR package and additionally performs the following data cleaning steps:
\itemize{
\item Copy values of AML into AMP, and CTX into CRO to simplify data
\item Converts telepath antibiotic and organism codes into full names
\item Removes suppression lowercase "s" from antibiotic columns, if present
\item Converts antibiotic columns to AMR::RSI format
\item Creates AMR_org using AMR::as.mo to facilitate further processing
\item Filters rows based on first isolate rules (see AMR::first_isolate)
\item Applies full EUCAST rules
}
}
\examples{
db <- RODBC::odbcConnectAccess2007('list_of_lists_db.accdb')
oxa <- pull_data(db, where_clause = "com_code Like '\%OXA\%'", limit = 0)
odbcClose(db)
oxa_clean <- clean_up(oxa)
}
\keyword{clean_up,}
\keyword{cleaning}
\keyword{data}
