% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_abx.R
\name{convert_abx}
\alias{convert_abx}
\title{Convert antibiotic codes generated from LCL telepath to full name (and vice-versa)}
\usage{
convert_abx(.data, .clean_up = F, .abbreviate = F, .been_cleaned = F)
}
\arguments{
\item{.data}{A character vector of antibiotic abbreviations or names}

\item{.clean_up}{Clean up and convert ambiguous names for use with AMR package:
\itemize{
\item 'Amp/Amoxil' = 'Ampicillin',
\item "Eryth/Clarith." = 'Erythromycin',
\item Ceftolozane-tazobactam' = 'Ceftolozane/tazobactam',
\item Caz/Avi' = 'Ceftazidime/avibactam',
\item Pip/Tazo' = 'Piperacillin/Tazobactam'
}}

\item{.abbreviate}{Converts a character vector of full antibiotic names to their LCL abbreviations}

\item{.been_cleaned}{Set as TRUE if the antibiotic names have been "cleaned up" with .clean_up}
}
\description{
Convert antibiotic codes generated from LCL telepath to full name (and vice-versa)
}
\examples{
abx_list <- c('AML', 'CIP', 'P/T')
convert_abx(abx_list)
[1] "Amp/Amoxil"    "Ciprofloxacin" "Pip/Tazo"

convert_abx(abx_list, .clean_up = T)
[1] "Ampicillin"              "Ciprofloxacin"           "Piperacillin/Tazobactam"
}
\keyword{abx,}
\keyword{convert_abx}
