% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_orgs.R
\name{convert_org}
\alias{convert_org}
\title{Convert organism codes generated from LCL telepath to full name (and vice-versa)}
\usage{
convert_org(.data, .clean_up = F, .abbreviate = F)
}
\arguments{
\item{.data}{A character vector of antibiotic abbreviations (e.g. c('HAIN', 'STAU')}

\item{.clean_up}{Clean up and convert ambiguous names for use with AMR package:
\itemize{
\item Diptheroids converted to Corynebacterium
\item Lactose fermenting coliform, Mixed coliforms, Coliform converted to E. coli
\item Mixed anaerobes converted to Bacteroides
\item Vancomycin resistant Entero- converted to VRE
\item sp. removed
\item Beta-haemolytic removed from Group X Strep
\item Alpha haemolytic Strep converted to Viridans Streptococcus
\item Anaerobes converted to Bacteroides
}}

\item{.abbreviate}{Converts a character vector of full organism names to their LCL abbreviations}
}
\description{
Convert organism codes generated from LCL telepath to full name (and vice-versa)
}
\examples{
org_list <- c('HAIN', 'STAU', 'HSB')
convert_org(org_list)
[1] "Haemophilus influenzae"                "Staphylococcus aureus"
[3] "Group B Beta haemolytic streptococcus"

convert_org(org_list, .clean_up = T)
[1] "Haemophilus influenzae" "Staphylococcus aureus"  "Group B Streptococcus"
}
\keyword{convert_org}
\keyword{org,}
