% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cpe_flextable.R
\name{make_cpe_flextable}
\alias{make_cpe_flextable}
\title{Create CPE summary table formatted using Flextable package}
\usage{
make_cpe_flextable(
  x,
  drop_abx = NULL,
  genotypes = list("OXA48", "KPC", "NDM", "VIM", "IMP")
)
}
\arguments{
\item{x}{A data frame or tibble in the format created using LCL::make_cpe_table}

\item{drop_abx}{If specified, a character vector of antibiotics to drop. Default NULL.}

\item{genotypes}{A list of genotypes (string format) of interest. Defaults to the big five: OXA48, KPC, NDM, VIM, IMP. Ensure that this matches the genotypes in the data input.}
}
\description{
This function takes data imported using LCL::pull_data that has undergone the following cleanup steps:
LCL::clean_up
LCL::make_cpe_table
}
\examples{
db <- RODBC::odbcConnectAccess2007('list_of_lists_db.accdb')
cpe <- pull_data(db, where_clause = "com_code Like '\%OXA\%' OR com_code Like '\%KPC\%'", limit = 0)
odbcClose(db)
cpe_clean <- clean_up(cpe)
cpe_table <- make_cpe_table(cpe_clean, genotypes = list('OXA48', 'KPC'))
cpe_flex_table <- make_cpe_flextable(cpe_table,
         drop_abx = c('Cefuroxime', 'Ceftriaxone', 'Doxycycline', 'Meropenem',
         Ertapenem', 'Teicoplanin', 'Vancomycin', 'Clindamycin', 'Erythromycin',
         Linezolid', 'Rifampicin', 'Fusidic acid'),
         genotypes = list('OXA48', 'KPC'))
}
\keyword{cleaning,}
\keyword{cpe}
\keyword{cpe,}
\keyword{data}
\keyword{table}
