% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cpe_table.R
\name{make_cpe_table}
\alias{make_cpe_table}
\title{Create CPE summary table}
\usage{
make_cpe_table(cpe_data, genotypes = list("OXA48", "KPC", "NDM", "VIM", "IMP"))
}
\arguments{
\item{cpe_data}{A data frame or tibble containing data pulled using LCL::pull_data and cleaned using LCL::clean_up}

\item{genotypes}{A list of genotypes (string format) of interest. Defaults to the big five: OXA48, KPC, NDM, VIM, IMP}
}
\description{
This function takes data imported using LCL::pull_data and cleaned using LCL::clean_up and
uses it to generate a CPE summary table.
}
\examples{
db <- RODBC::odbcConnectAccess2007('list_of_lists_db.accdb')
cpe <- pull_data(db, where_clause = "com_code Like '\%OXA\%' OR com_code Like '\%KPC\%'", limit = 0)
odbcClose(db)
cpe_clean <- clean_up(cpe)
cpe_table <- make_cpe_table(cpe_clean, genotypes = list('OXA48', 'KPC'))
}
\keyword{cleaning}
\keyword{cpe,}
\keyword{data}
