% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pull data from Access database in list-of-lists format}
\usage{
pull_data(
  database = NULL,
  where_clause = NULL,
  select_clause = NULL,
  limit = 10
)
}
\arguments{
\item{database}{An RODBC Access database object, remember to close after open}

\item{where_clause}{An SQL WHERE clause. Use single quotes for strings and # for dates}

\item{select_clause}{An SQL SELECT clause, which can be any of the Access database column names, e.g.
tPatients.num_hospital, tSamples.num_spec, date_collected, org_code, spec_type,
antibiotic codes e.g. MER, ETP}

\item{limit}{Limit the number of rows returned (useful to test large queries), defaults to 10. Use 0 to get all.}
}
\description{
Pull data from Access database in list-of-lists format
}
\examples{
db <- RODBC::odbcConnectAccess2007('list_of_lists_db.accdb')
oxa <- pull_data(db, where_clause = "com_code Like '\%OXA\%'", limit = 0)
odbcClose(db)
}
\keyword{access}
\keyword{data}
\keyword{gather,}
\keyword{pull_data,}
